+++ Assets/Scripts/GlobalValues/Alphabet.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 24ffee44b9cb62d4485574e7a3315883
+guid: 67a5d70f93ec4b14abaa78799fcb265a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
