+++ Assets/Scripts/GlobalValues/LetterQueue.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 67a5d70f93ec4b14abaa78799fcb265a
+guid: 24ffee44b9cb62d4485574e7a3315883
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
